// web/src/api/generated/model/passkey.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * stocat
 * OpenAPI spec version: 0.0.1
 */

export interface Passkey {
  created_at: string;
  id: string;
  /** @nullable */
  last_used_at: string | null;
  name: string;
  /** The authenticator backs up the passkey, for example to a cloud account. */
  synced: boolean;
}